refactor: cache
