+++ src/index.js
@@ -1,61 +1,8 @@
-import { camelize, escapeReg, isBoolean, isFunction, isArray } from './util'
+import { camelize, escapeReg, isBoolean } from './util'
 import { assert, warn } from './debug'
 import apiCreator from './creator'
 import instantiateComponent from './instantiate'
-
-const cache = {
-  instances: [],
-  add(component) {
-    let alreadyIn = false
-    const instances = cache.instances
-    const len = instances.length
-    for (let i = 0; i < len; i += 1) {
-      const ins = instances[i]
-      if (ins === component) {
-        alreadyIn = true
-        break 
-      }
-    }
-    if (!alreadyIn) {
-      instances.push(component)
-    }
-  },
-  remove(component) {
-    const instances = cache.instances
-    const len = instances.length
-    for (let i = 0; i < len; i += 1) {
-      const ins = instances[i]
-      if (ins === component) {
-        instances.splice(i, 1)
-        return
-      }
-    }
-  }
-}
-
-function batchDestroy(filter) {
-  const hasFilter = isFunction(filter)
-  const instancesCopy = cache.instances.slice()
-  const instances = hasFilter ? filter(instancesCopy) : instancesCopy
-  if (!isArray(instances)) {
-    return
-  }
-  if (hasFilter) {
-    instances.forEach(ins => {
-      if (ins && isFunction(ins.remove)) {
-        ins.remove()
-        cache.remove(ins)
-      }
-    })
-  } else {
-    instances.forEach(ins => {
-      if (ins && isFunction(ins.remove)) {
-        ins.remove()
-      }
-    })
-    cache.instances.length = 0
-  }
-}
+import { batchDestroy } from './cache'
 
 function install(Vue, options = {}) {
   const {componentPrefix = '', apiPrefix = '$create-'} = options
@@ -65,7 +12,7 @@ function install(Vue, options = {}) {
       single = events
       events = []
     }
-    const api = apiCreator.call(this, Component, events, single, cache)
+    const api = apiCreator.call(this, Component, events, single)
     const createName = processComponentName(Component, {
       componentPrefix,
       apiPrefix,
